requirements.txt
click==6.7
Flask==0.12.2
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.0
pymongo==3.6.1
python-ldap==2.5.2
PyYAML==3.12
six==1.11.0
Werkzeug==0.14.1
WTForms==2.1
